+++ src/bot/bot.ts
@@ -1,10 +1,10 @@
 import 'dotenv/config'
 import { pipe } from 'fp-ts/lib/function'
 import _ from 'lodash'
-import { getItems } from '../utils/feed.util'
-import { sendRelease } from '../utils/send.util'
-import { storeNewFeedItem } from '../utils/store.util'
-import { validateFeed } from '../utils/validate.util'
+import { getItems } from '../utils/feed.utils'
+import { sendRelease } from '../utils/send.utils'
+import { storeNewFeedItem } from '../utils/store.utils'
+import { validateFeed } from '../utils/validate.utils'
 
 export const startBot = () => pipe(
     getItems(),
